refactor(FinalWinning): 수익을 계산하는 책임을 가짐
- 이전: Participants 가 수익을 계산하는 책임을 가짐
- 이후: FinalWinning 가 수익을 계산하는 책임을 가짐
